+++ standard/CHANGELOG.md
@@ -5,6 +5,7 @@
 - FIX: With custom slotHeaderFormats, navLinks sometimes disabled (#8064)
 - FIX: In TimeGrid view, when snapDuration smaller than slotDuration, incorrect date-selection coordinates
 - FIX: In React/Preact, expose the CalendarRef type
+- FEATURE: Warn on invalid class-name inputs
 
 
 ## v7.0.0-rc.2